# Benchmark cel.UnstructuredToVal