… staging/src/k8s.io/apiextensions-apiserver/third_party/forked/celopenapi/model/escaping.go
… staging/src/k8s.io/apiextensions-apiserver/third_party/forked/celopenapi/model/escaping_test.go