<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?
/kind cleanup
<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->

#### What this PR does / why we need it:

Given that the escapeKeyProps() function spends a significant percentage of its time inside the regex.ReplaceAllStringFunc() function, a potential optimization strategy is to avoid that expensive function call. An assumption is that the vast majority of map keys we encounter in the wild do not contain non-alphanumeric characters, in which case we can potentially avoid the regular expression matching operation.

The result from benchmark test:
**Before**
```
BenchmarkUnstructuredToVal
BenchmarkUnstructuredToVal-4   	  593342	      1696 ns/op	    1248 B/op	      14 allocs/op
PASS

BenchmarkUnstructuredToValWithEscape
BenchmarkUnstructuredToValWithEscape-4   	  826821	      1326 ns/op	     892 B/op	      12 allocs/op
PASS
```
**After**
```
BenchmarkUnstructuredToVal
BenchmarkUnstructuredToVal-4   	 1078765	      1058 ns/op	    1208 B/op	      11 allocs/op
PASS

BenchmarkUnstructuredToValWithEscape
BenchmarkUnstructuredToValWithEscape-4   	 1473859	       790.9 ns/op	     856 B/op	       9 allocs/op
PASS
```
The alternative solutions:
1. Set of Characters
```
BenchmarkUnstructuredToVal
BenchmarkUnstructuredToVal-4   	  895012	      1133 ns/op	    1208 B/op	      11 allocs/op
PASS

BenchmarkUnstructuredToValWithEscape-4   	 1000000	      1014 ns/op	     856 B/op	       9 allocs/op
PASS
```
2. Hardcode conditions
```
BenchmarkUnstructuredToVal
BenchmarkUnstructuredToVal-4   	 1025188	      1580 ns/op	    1208 B/op	      11 allocs/op
PASS

BenchmarkUnstructuredToValWithEscape
BenchmarkUnstructuredToValWithEscape-4   	 1467898	       787.9 ns/op	     856 B/op	       9 allocs/op
PASS
```

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```