@cici37 2022-05-10
@liggitt Would you please have a look when have time? It is a performance improvement previously pending because of code freeze :) Thank you