@jpbetz 2022-05-11
/lgtm

I think there is an opportunity here to write a benchmark that more clearly highlights the impact of this change.  I wonder if the use of short (1-2 char) identifiers might not fully highlight some of the performance benefits of this change?